+++ src/System/Taffybar/Widget/Text/CPUMonitor.hs
@@ -5,7 +5,6 @@ import qualified Text.StringTemplate as ST
 import System.Taffybar.Information.CPU
 import System.Taffybar.Widget.Generic.PollingLabel ( pollingLabelNew )
 import qualified GI.Gtk
-import qualified Data.Text as T
 
 -- | Creates a simple textual CPU monitor. It updates once every polling
 -- period (in seconds).